Screen Logic Implemented
(1) There are now four screens:
Splash for the splash image
Start for the start menu
Game for the game
ExitScreen for win or loss screen

The splash screen stays up for an arbitrary amount of time that we can
set by clicking the script component of the camera.
Then the start screen loads, and we can click new game to start a game,
help for info (not yet implemented), and exit to....exit.

(2) I've also attached an Ultimate_Solitaire c# to the main camera in
the game scene, so we can use that for the game logic if we need to.