Created end screen, tidied up help text
The end screen is now fully functional. You just have to load it and
call the win or lose screen depending on whether you've won or lost
(once the game is working we can add the function call.)
In addition, the three buttons will either reload the game, go to main
menu, or quit.

I also formatted the help text and also made it more succinct.