Lots of updates
Working on adding a method so that when an object leaves a trigger, it
reverses any move changes that may have been done. That way, if you
decide you don't want to move somewhere, you can choose not to.

Also much more layer work. Layers for cards seem to work except for when
a foundation throws a null reference exception - this throws everything
off.