Added MultiReset() and changed some of the logic related to OnMouseDown()
(1) Added MultiReset(). Now when you release the mouse button and the
move isn't valid but you were moving a stack, the game will attempt to
return them to their original positions and layers.

(2) Moved the logic for OnMouseDown card states out of OnMouseDown into
separate functions. Promoting OOP and readability, as well as speed of
game.

There are still a couple of layering bugs with moving multiple cards as
well as when you release the mouse button but I believe these will be
easily fixable now that the internal logic is shaping up.