Scoring and calling win / lose framework
Two testing buttons added to game screen to test win / lose.

Game uses persistent object to store score and bool win / loss. When the
scene switches, the object survives and is used to determine whether to
show the win or lose screen as well as pulling the score, and updating
the high score if needed.